Refactor by creating a features folder
- Create a updateTranslationFileOnCommit file in this new folder
